requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.4
djangorestframework==3.14.0
easyocr==1.7.0
filelock==3.12.2
google==3.0.0
google-cloud==0.34.0
idna==3.4
imageio==2.31.1
Jinja2==3.1.2
lazy_loader==0.3
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.1
ninja==1.11.1
numpy==1.25.2
opencv-python-headless==4.8.0.76
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
pyclipper==1.3.0.post4
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0.1
requests==2.31.0
scikit-image==0.21.0
scipy==1.11.1
shapely==2.0.1
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
sympy==1.12
tifffile==2023.8.12
torch==2.0.1
torchvision==0.15.2
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
